remove fucking defines base code, refractoring and split files
