+++ runtime/ishtar.vm/runtime/gc/GCLayout_Debug.cs
@@ -1,0 +1,7 @@
+namespace ishtar.runtime.gc;
+
+public unsafe interface GCLayout_Debug
+{
+    public void find_leak();
+    public void dump(string file);
+}
