+++ runtime/ishtar.vm/runtime/gc/GcHeapUsageStat.cs
@@ -1,0 +1,10 @@
+namespace ishtar.runtime.gc;
+
+public struct GcHeapUsageStat
+{
+    public long pheap_size;
+    public long pfree_bytes;
+    public long punmapped_bytes;
+    public long pbytes_since_gc;
+    public long ptotal_bytes;
+}
